Add an error for global uniform parameter declarations
A global uniform parameter in HLSL might canonically be defined like this:

```hlsl
uniform float gSomeParameter;
```

The fxc and dxc compilers automatically collect all such parameters into a synthesized constant buffer, along the lines of:

```hlsl
cbuffer $Globals
{
    float gSomeParameter;
}
```

Slang currently supports parsing and semantic checking of declarations like the above, and computes shader parameter layout/binding information that is appropriate for a constant buffer like `$Globals` above, but it does not include the support to emit HLSL or GLSL code that matches that layout, so that use of global uniforms in Slang is silently unsupported.

Making this problem worse, the HLSL language is quite lax, and will parse the following as shader parameters as well:

```hlsl
int gCounter = 0;
const float kScaleFactor = 2.0f;
```

Each of those declarations introduces a global shader parameter, and then provides a default value for it via the initializer. These declarations do *not* introduce an ordinary global variable or constant as might be expected.

(For anybody who wants to know, `static` is required to introduce a "real" global variable (although it will be a *thread-local* global in practice), while `static const` is required to introduce a global constant)

I was not too worried about users trying to use global-scope uniforms and failing (since that has fallen out of common HLSL/GLSL practice), but the possibility that users might try to declare global variables/constants and get shader parameters by mistake creates more of a risk so that this hole is worth plugging.

The right long-term fix is of course to support the intended semantics of global-scope uniforms, but that feature needs to be prioritized against other requests.

A few of the Slang tests were unwittingly relying on this functionality, including some compute tests that seemingly got away with it based on the DXBC generated from the HLSL output by Slang just happening to match the layout they expected. These tests have all been tweaked to use explicit `cbuffer`s or `ParameterBlock`s instead.