Fix checking of interface conformances for nested types
Before this change, code like the following would crash the compiler:

```hlsl
interface IThing { /* ... */ }
struct Outer
{
    struct Inner : IThing
    {}
}
/* go on to use Outer.Inner */
```

The problem was that the front-end logic for checking interface conformances was *only* checking declarations at the top level of a module, or nested under a generic.
This change fixes the logic to recurse through the entire tree of declarations.
I have added a test case that uses a nested `struct` type to satisfy an associated type requirement, to confirm that the new check works as intended.