Produce a better error message on errors in inheritance clauses (#828)
If I write:

```
struct Foo : IDoesntExist
{}
```

Then currently Slang complains first that `IDoesntExist` is an undefined identifier, then it complains that it expected an interface type in the inheritance clause. The second ("cascading") error isn't really helpful, because of *course* if something isn't defined it isn't an interface.

This change detects the case where the type expression is erroneous so that we avoid the cascading error.